requirements/base.txt
python-slugify==8.0.4  # https://github.com/un33k/python-slugify
Pillow==10.4.0  # https://github.com/python-pillow/Pillow
argon2-cffi==23.1.0  # https://github.com/hynek/argon2_cffi
whitenoise==6.7.0  # https://github.com/evansd/whitenoise
redis==5.0.7  # https://github.com/redis/redis-py
hiredis==3.0.0  # https://github.com/redis/hiredis-py
celery==5.4.0  # pyup: < 6.0  # https://github.com/celery/celery
django-celery-beat==2.6.0  # https://github.com/celery/django-celery-beat

# Django
# ------------------------------------------------------------------------------
django==5.0.0  # pyup: < 5.0  # https://www.djangoproject.com/
django-environ==0.11.2  # https://github.com/joke2k/django-environ
django-model-utils==4.5.1  # https://github.com/jazzband/django-model-utils
django-redis==5.4.0  # https://github.com/jazzband/django-redis
# Django REST Framework
djangorestframework==3.15.2  # https://github.com/encode/django-rest-framework
django-cors-headers==4.4.0  # https://github.com/adamchainz/django-cors-headers
# DRF-spectacular for api documentation
drf-spectacular==0.27.2  # https://github.com/tfranzel/drf-spectacular

#installed apps by me
djangorestframework-simplejwt==5.3.1
django-jazzmin==2.0.0
django-modeltranslation==0.18.11
django-phonenumber-field==8.0.0
django-filter==24.2
phonenumbers==8.13.40
channels[daphne]==4.0.0
channels-redis==4.2.0
djangochannelsrestframework==1.2.0
boto3==1.35.2
django-storages==1.14.4
twisted[tls,http2]==24.7.0
daphne==4.1.2 
djangorestframework-gis==1.1
django-minio-storage==0.5.7 
drf-writable-nested==0.7.1 
reportlab==4.2.5 
django-import-export==4.3.3 
drf-writable-nested==0.7.1 
python-docx==1.1.2 
qrcode[pil]==8.0

